Migrate to typescript partially
