+++ components/atoms/Heading/type.tsx
@@ -1,0 +1,6 @@
+import { ReactNode } from "react";
+
+export type Props = {
+  children?: ReactNode;
+  style?: string;
+};
