--- components/blog/atoms/Heading/H1.jsx
@@ -1,9 +1,0 @@
-import React from 'react';
-
-function H1({children, style}) {
-  return (
-    <h1 className={`text-2xl font-bold ${style}`}>{children}</h1>
-  );
-}
-
-export default H1;
