--- components/blog/atoms/Href/Href.jsx
@@ -1,20 +1,0 @@
-import React from "react";
-import Link from "next/link";
-
-function Href({ children, slug, originalUrl }) {
-  return (
-    <>
-      {slug ? (
-        <Link href={slug}>
-          <div className="cursor-pointer">{children}</div>
-        </Link>
-      ) : (
-        <a href={`${originalUrl}`} target="_blank" rel="noreferrer noopener">
-          {children}
-        </a>
-      )}
-    </>
-  );
-}
-
-export default Href;
