--- components/templates/Container.jsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-function Container({children}) {
-  return (
-    <div className="container mx-auto px-6">
-      {children}
-    </div>
-  );
-}
-
-export default Container;
