+++ components/templates/index.tsx
