--- jsconfig.json
@@ -1,10 +1,0 @@
-{
-  "compilerOptions": {
-    "baseUrl": ".",
-    "paths": {
-      "@/blog/components/*": ["components/blog/*"],
-      "@/utils/*": ["utils/*"],
-      "@/services/*": ["services/*"],
-    }
-  }
-}
