--- pages/_app.js
@@ -1,31 +1,0 @@
-import React, { } from 'react';
-import 'tailwindcss/tailwind.css'
-import '../styles/global.scss'
-
-import { withRouter } from 'next/router'
-import { BlogLayout, MainLayout } from 'components/templates/'
-
-const layouts = {
-  main: MainLayout,
-  blog: BlogLayout,
-}
-
-function MyApp({ Component, pageProps, router }) {
-  let Layout;
-  switch (router.pathname.split("/")[1]) {
-    case "blog":
-      Layout = layouts['blog']
-      break;
-  
-    default:
-      Layout = layouts['main']
-      break;
-  }
-  return (
-    <Layout>
-      <Component {...pageProps} />
-    </Layout>
-  )
-}
-
-export default withRouter(MyApp)
