+++ pages/blog/post/[slug].jsx
@@ -1,8 +1,8 @@
 import React, { useEffect, useState } from "react";
 // import { ArticleDetail, Loader } from "../../components";
-import { Loader } from "@/blog/components/atoms";
-import { ArticleDetail } from "@/blog/components/organisms";
-import { getPostDetails, getPosts } from "services/";
+import { Loader } from "../../../components/atoms";
+import { ArticleDetail } from "../../../components/organisms";
+import { getPostDetails, getPosts } from "../../../services";
 import { useRouter } from "next/router";
 import Head from "next/head"
